show on index coming events

+++ Mosby/src/main/java/com/mosby/view/web/servlets/EventPageServlet.java
@@ -24,7 +24,7 @@ public class EventPageServlet extends HttpServlet {
         if (request.getParameter("eventId") != null) {
             int eventId = Integer.parseInt(request.getParameter("eventId"));
             Event event = new ReadGenericObjectService<Event>((Class<Event>) Event.class).readById(eventId); 
-            List <TicketInfo> ticketsInfoList = new ReadGenericObjectService<TicketInfo>((Class<TicketInfo>) TicketInfo.class).readListByField("event_ref", (Integer)eventId);
+            List <TicketInfo> ticketsInfoList = new ReadGenericObjectService<TicketInfo>((Class<TicketInfo>) TicketInfo.class).readListByField("event_ref", eventId);
             request.setAttribute("event", event);
             request.setAttribute("tickets", ticketsInfoList);
             request.getRequestDispatcher("/pages/eventPage.jsp").forward(request, response);
